Changes for readr 2.0.0
In previous version of readr the reading functions took literal data as any
filename with a newline in it or vectors of length > 1. In readr 2.0.0 vectors
of length > 1 are now  assumed to correspond to multiple files. Because of this
we now use a more explicit way to represent literal data, by using the `I()`
function around the input.

In addition the way that the problems attribute is stored has changed
and is now an external pointer. Because of this, two datasets will never
be identical, so one test needed to be changed to expect_equal instead.